% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertDf2Matrix}
\alias{convertDf2Matrix}
\title{Convert an RSA data frame to matrix for running RSA}
\usage{
convertDf2Matrix(df)
}
\arguments{
\item{df, }{data frame containing at least group, quantity and two more cols (for items)}
}
\value{
, fill this out
}
\description{
Convert an RSA data frame to matrix for running RSA
}
\examples{
d <- data.frame(scales = rep("some_all", 10),
stars = as.factor(rep(1:5, 2)),
degrees = c(rep("strong", 5), rep("weak", 5)),
speaker.p = c(0, 0, 0, 0.3, 0.7,
0, 0.1, 0.15, 0.35, 0.40),
pragmatics = c(0, 0, 0, 0.15, 0.85,
0, 0.1, 0.25, 0.5, 0.15))
cData <- convertData(d)
convertDf2Matrix(cData)

}
\keyword{data_org}

