% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convertMatrix2Df}
\alias{convertMatrix2Df}
\title{Convert an RSA matrix to data frame for return to user
append group and quantity columns (for matching in return df)}
\usage{
convertMatrix2Df(m, group)
}
\arguments{
\item{m, }{matrix}
}
\value{
, fill this out
}
\description{
Convert an RSA matrix to data frame for return to user
append group and quantity columns (for matching in return df)
}
\examples{
d <- data.frame(scales = rep("some_all", 10),
stars = as.factor(rep(1:5, 2)),
degrees = c(rep("strong", 5), rep("weak", 5)),
speaker.p = c(0, 0, 0, 0.3, 0.7,
0, 0.1, 0.15, 0.35, 0.40),
pragmatics = c(0, 0, 0, 0.15, 0.85,
0, 0.1, 0.25, 0.5, 0.15))
cData <- convertData(d)
mData <- convertDf2Matrix(cData$runData)
convertMatrix2Df(mData)

}
\keyword{data_org}

