% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_pragmods.R
\docType{data}
\name{d_pragmods}
\alias{d_pragmods}
\title{Data from Rational speech act models of pragmatic reasoning in reference games}
\description{
This data set contains ....
}
\details{
1) quantityVarName :: "stars"

2) semanticsVarName :: "speaker.p"

3) itemVarName :: "words"

4) groupVarName :: "scale"

other) experiment number :: "exp"

other) Pragmatic judgments study 1 (e6), study 2 (e11)
}
\references{
\url{https://github.com/langcog/scalar_implicature}
}
\keyword{data}

