% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{informativity}
\alias{informativity}
\title{RSA informativity}
\usage{
informativity(m_u, alpha = 1, cost = 0)
}
\arguments{
\item{m_u, }{literal semantics of meaning given utterance}

\item{alpha, }{decision noise parameter (speaker's deviation
from optimal action selection)}

\item{cost, }{cost of utterance u}
}
\value{
, fill this out
}
\description{
e^(-alpha * (-log(p(m|u)) - cost))
}
\examples{
informativity(0.5, 1, 0) == 0.5
informatitivy(0, 0, 0) == 0.0

}
\keyword{surprisal}

