% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{normVec}
\alias{normVec}
\title{Normalize vectors}
\usage{
normVec(v)
}
\arguments{
\item{v, }{vector to be normalized}
}
\value{
, fill this out
}
\description{
Return a normalized vector
}
\examples{
vec1 <- c(1, 1, 1)
norm(vec1)
vec2 <- c(0, 0, 0, 0)
norm(vec2)

}
\keyword{normalization}

