% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_2alts.R
\docType{data}
\name{peloquinFrank_2alts}
\alias{peloquinFrank_2alts}
\title{Two alts model data from Peloquin\& Frank (2016)}
\usage{
data(peloquinFrank_2alts)
}
\description{
Two alts model data from Peloquin\& Frank (2016)
}
\keyword{data}

