% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_4alts.R
\docType{data}
\name{peloquinFrank_4alts}
\alias{peloquinFrank_4alts}
\title{Four alts model data from Peloquin\& Frank (2016)}
\usage{
data(peloquinFrank_4alts)
}
\description{
Four alts model data from Peloquin\& Frank (2016)
}
\keyword{data}

