% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_5alts.R
\docType{data}
\name{peloquinFrank_5alts}
\alias{peloquinFrank_5alts}
\title{Five alts model data from Peloquin\& Frank (2016)}
\usage{
data(peloquinFrank_5alts)
}
\description{
Five alts model data from Peloquin\& Frank (2016)
}
\keyword{data}

