% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotFns.R
\name{plotSemantics}
\alias{plotSemantics}
\title{Plot literal listener semantics
Compare dimensions of two matrices or vectors}
\usage{
plotSemantics(processedData)
}
\arguments{
\item{processedData, }{data frame after proceessing (from calling processData())}

\item{m1, }{matrix 1}
}
\value{
, fill this out
}
\description{
Plot literal listener semantics
Compare dimensions of two matrices or vectors
}
\examples{
d <- data(practiceData)
processedData <- processData(d)
plotSemantics(processedData)

}
\keyword{plotting}

