% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/practiceData.R
\docType{data}
\name{practiceData}
\alias{practiceData}
\title{Practice data from Peloquin \& Frank (2015)}
\usage{
data(practiceData)
}
\description{
A test data set of four variables \code{scale}, \code{degree},
\code{stars}, \code{semantics}
}
\keyword{data}

