% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{processData}
\alias{processData}
\title{Process data to correct structure for conversion to matrix for reasoning()}
\usage{
processData(data, group, quantity, item, semantics)
}
\arguments{
\item{data, }{data for running rsa}

\item{group, }{grouping variable (e.g. scales <some_all>)}

\item{quantity, }{quantity for assessing semantic compatibility (e.g. stars <1...5>)}

\item{item, }{invidiual word or degree level items (e.g. <strong, weak> or <hi, low> or <some, all>)}

\item{semantics, }{normalized distribution over quantity (e.g. compatibility DV)}
}
\value{
, fill this out
}
\description{
Data should be passed in 'long' format with 4 required fields
1) Group ::
2) Quantity ::
3) Item ::
4) Semantics ::
}
\examples{
d <- data.frame(scales = rep("some_all", 10),
stars = rep(1:5, 2),
degrees = c(rep("strong", 5), rep("weak", 5)),
speaker.p = c(0, 0, 0, 0.3, 0.7,
0, 0.1, 0.15, 0.35, 0.40),
listener.p = c(0, 0, 0, 0.15, 0.85,
0, 0.1, 0.25, 0.5, 0.15))
convertData(d, group = "scales", quantity = "stars", item = "degrees", semantics = "speaker.p")

}
\keyword{data_org}

