% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{recurse}
\alias{recurse}
\title{Run one full recursion between listener -> speaker -> listener}
\usage{
recurse(m, costs = m - m, priors = rep(1, nrow(m)), alpha = 1)
}
\arguments{
\item{m, }{matrix of semantics (rows = meaning, cols = words)}

\item{costs, }{matrix of costs (default is 0)}

\item{priors, }{default uniform, vector of length nrow() [semantic quantity]}

\item{alpha, }{decision hyper-param}
}
\value{
, fill this out
}
\description{
Run one full recursion between listener -> speaker -> listener
}
\examples{
m <- matrix(data = c(0, 0.2, 0.25, 0.25, 0.3, 0, 0, 0, 0.3, 0.7), nrow = 5)
rownames(m) <- 1:5
colnames(m) <- c("item1", "item2")
recurse(m)
recurse(recurse(m))
costs <- matrix(data = c(rep(-0.2, 5), rep(-0.8, 5)), nrow = 5)
recurse(m, costs)
recurse(recurse(m, costs), costs)

}
\keyword{recursion}

