% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{renameRSACols}
\alias{renameRSACols}
\title{Rename group, quantity, item and semantics columns in user df for use in RSA}
\usage{
renameRSACols(df, group, quantity, item, semantics)
}
\arguments{
\item{df, }{data frame of measurements}

\item{group, }{group name (e.g. "scales")}

\item{quantity, }{quantity we're quantifying over (e.g. "stars")}

\item{item, }{items we're examining (either individual words
("some", "all", or degrees, "weak", "strong))}

\item{semantics, }{normalized compatibility measures}
}
\value{
, fill this out
}
\description{
Rename group, quantity, item and semantics columns in user df for use in RSA
}
\examples{
d <- data.frame(scales = rep("some_all", 10),
stars = as.factor(rep(1:5, 2)),
degrees = c(rep("strong", 5), rep("weak", 5)),
speaker.p = c(0, 0, 0, 0.3, 0.7,
0, 0.1, 0.15, 0.35, 0.40),
pragmatics = c(0, 0, 0, 0.15, 0.85,
0, 0.1, 0.25, 0.5, 0.15))
newDf <- renameRSACols(d, group = "scales", quantity = "stars",
item = "degrees", semantics = "speaker.p")
newDf$data
newDf$labels

}
\seealso{
Called by \code{processData()}
}
\keyword{data_org}

