% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rrrsa.R
\docType{package}
\name{rrrsa}
\alias{rrrsa}
\alias{rrrsa-package}
\title{rrrsa package for running Rational speech-act models}
\description{
for functions \code{link{reason}} and \code{link{informativity}}.
}
\references{
Frank \& Goodman (2012)
\url{http://langcog.stanford.edu/papers_new/frank-2012-science.pdf}
}
\seealso{
\code{\link{reason}}

\code{link{informativity}}
}

