% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.batchRun}
\alias{rsa.batchRun}
\title{Run RSA model on processed data}
\usage{
rsa.batchRun(data, alpha = 1, depth = 1)
}
\arguments{
\item{data, }{data passed after call to \code{processData()}}

\item{alpha, }{alpha parameter}

\item{depth, }{depth of recursion}
}
\value{
, fill this out
}
\description{
Only accepts processed data as from processData()
}
\examples{
print("not yet implemented")

}
\keyword{main}

