% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.bestFit}
\alias{rsa.bestFit}
\title{Wrapper to tunDepthAlpha}
\usage{
rsa.bestFit(data, quantityVarName, semanticsVarName, itemVarName,
  groupName = NA, compareDataName, costsVarName = NA, priorsVarName = NA,
  depths = 1, alphas = 1, compareIndices = NA)
}
\arguments{
\item{data, }{tidied data}

\item{quantityVarName, }{entity name we're quantifying over}

\item{semanticsVarName, }{semantic values for inference computation}

\item{itemVarName, }{unique items were comparing, probaby words}

\item{groupName, }{grouping variable if we have one}

\item{compareDataName, }{pragmatic judgments we're comparing to}

\item{costsVarName, }{costs variable name}

\item{priorsVarName, }{priors variable name}

\item{depths, }{vector of depths (in integers) for tuning}

\item{alphas, }{vector of alphas for tuning}

\item{compareIndices, }{specific indices in data frame}
}
\value{
list of length(alphas) * length(depths) tuples with (correlation, depth, alpha)
}
\description{
Return best model run (correlation, depth, alpha) tuple
each element includes a tuple of (correlation, alpha, depth).
}
\examples{
d <- peloquinFrank_5Alts
alphas <- seq(1, 3, by = 0.1)
depths <- 1:3
checkWords <- c("some", "all", "good", "excellent", "liked", "loved", "memorable", "unforgettable",
"palatable", "delicious")
compareIndices <- which(peloquinFrank_5Alts$words \%in\% checkWords)
bestModel <- rsa.bestFit(data = d, groupName = "scale",
quantityVarName = "stars", itemVarName = "words",
semanticsVarName = "speaker.p", compareDataName = "e11",
compareIndices = compareIndices, alphas = alphas, depths = depths)

}
\keyword{data}
\keyword{tuning}

