% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.fullRecursion}
\alias{rsa.fullRecursion}
\title{Run one full recursion between listener -> speaker -> listener}
\usage{
rsa.fullRecursion(m, costs = rep(0, ncol(m)), priors = rep(1, nrow(m)),
  alpha = 1)
}
\arguments{
\item{m, }{matrix of semantics (rows = meaning (m rows), cols = words (n cols))}

\item{costs, }{length m vector of costs (default is 0 valued vector)}

\item{priors, }{default uniform, vector of length nrow() [semantic quantity]}

\item{alpha, }{decision hyper-param}
}
\value{
matrix of posterior values
}
\description{
Run one full recursion between listener -> speaker -> listener
}
\examples{
m <- matrix(data = c(0, 0.2, 0.25, 0.25, 0.3, 0, 0, 0, 0.3, 0.7), nrow = 5)
rownames(m) <- 1:5
colnames(m) <- c("item1", "item2")
rsa.fullRecursion(m)
rsa.fullRecursion(rsa.fullRecursion(m))

}
\keyword{recursion}

