% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.informativity}
\alias{rsa.informativity}
\title{RSA informativity}
\usage{
rsa.informativity(m_u, alpha = 1, cost = 0)
}
\arguments{
\item{m_u, }{literal semantics of meaning given utterance}

\item{alpha, }{decision noise parameter (speaker's deviation
from optimal action selection)}

\item{cost, }{cost of utterance u}
}
\value{
, inormativity of utterance u given meaning m and cost
}
\description{
Calculate: e^(-alpha * (-log(p(m|u)) - cost))
}
\examples{
rsa.informativity(m_u = 0.5, alpha = 1, cost = 0) == 0.5
rsa.informatitivy(m_u = 0, alpha = 1, cost = 0) == 0.0
rsa.informatitivy(m_u = 0.5, alpha = 1, cost = 2)

}
\keyword{surprisal}

