% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.normVec}
\alias{rsa.normVec}
\title{Normalize vectors}
\usage{
rsa.normVec(v)
}
\arguments{
\item{v, }{vector to be normalized}
}
\value{
, normalized vector
}
\description{
Return a normalized vector
}
\examples{
vec1 <- c(1, 1, 1)
normVec(vec1)
vec2 <- c(0, 0, 0, 0)
normVec(vec2)

}
\keyword{data}
\keyword{processing}

