% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.reason}
\alias{rsa.reason}
\title{Run (multiple) iterations RSA}
\usage{
rsa.reason(m, costs = rep(0, ncol(m)), priors = rep(1, nrow(m)),
  depth = 1, alpha = 1, usePriorEveryRecurse = TRUE)
}
\arguments{
\item{m, }{matrix of semantics with 'items' cols and 'quantity' rows}

\item{costs, }{ncol(m) vector of costs (default is 0 valued vector)}

\item{priors, }{nrow(m) vector of priors (default is uniform)}

\item{depth, }{number of recursions}

\item{alpha, }{decision hyper-parameter}

\item{usePriorEveryRecurse, }{boolean incorporate priors during each full recursion}
}
\value{
matrix of posterior values
}
\description{
Return matrix after undergoing number 'depth' recursions
}
\examples{
m <- matrix(data = c(0, 0.2, 0.25, 0.25, 0.3, 0, 0, 0, 0.3, 0.7), nrow = 5)
rownames(m) <- 1:5
colnames(m) <- c("item1", "item2")
rsa.reason(m, depth = 0)
rsa.reason(m, depth = 2)

}
\keyword{recursion}

