% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.renameCol}
\alias{rsa.renameCol}
\title{Rename df columns}
\usage{
rsa.renameCol(df, currNames, replacements)
}
\arguments{
\item{df, }{data frame to change col names}

\item{currNames, }{names we want to replace (passed as characters)}

\item{replacements, }{values used as replacements}
}
\value{
data frame with changed col names
}
\description{
return a data frame with columns renamed
(if they exist in the df passed in)
}
\examples{
df <- data.frame(LOWER_CASE = letters[1:3], upper_case = LETTERS[1:3])
rsa.renameCol(df, c("LOWER_CASE", "upper_case"), c("lower_case", "UPPER_CASE"))

}
\keyword{data}
\keyword{processing}

