% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.runDf}
\alias{rsa.runDf}
\title{Run RSA on a tidied data frame (assumes one group)}
\usage{
rsa.runDf(data, quantityVarName, semanticsVarName, itemVarName,
  costsVarName = NA, priorsVarName = NA, depth = 1, alpha = 1)
}
\arguments{
\item{data, }{tidied data}

\item{quantityVarName, }{entity name we're quantifying over
(i.e. "stars" in Peloquin & Frank (2016))}

\item{semanticsVarName, }{literal listener semantics for RSA computations}

\item{itemVarName, }{unique items were comparing, probaby words
(i.e. "degrees" in Peloquin & Frank (2016))}

\item{costsVarName, }{costs variable name}

\item{priorsVarName, }{priors variable name}

\item{depth, }{recursive depth parameter}

\item{alpha, }{decision noise parameter level}
}
\value{
data frame with posterior predictions 'preds' column appended
}
\description{
Expects tidied data with three required and two optional fields
1) quanityVarName         :: entity name we're quantifying over
2) itemVarName            :: unique items were considering,
3) semanticsVarName       :: literal listener semantics
4) optional costsVarName  :: costs
5) optional priorsVarName :: priors
rsa-ready, tidied data should have a unique semantic value for each quantity * item combination.
(see formatting of peloquinFrank_2Alts data).
}
\examples{
rsa.runDf(peloquinFrank_2Alts, "stars", "speaker.p", "words")

}
\keyword{functionality}
\keyword{primary}
\keyword{run}

