% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.runDf_grouped}
\alias{rsa.runDf_grouped}
\title{Run RSA on tidied data frame with multiple groups
rsa-ready, tidied data should have a unique semantic value for each quantity * item combination.
(see formatting of peloquinFrank_2Alts data).}
\usage{
rsa.runDf_grouped(data, groupVarName, quantityVarName, semanticsVarName,
  itemVarName, costsVarName = NA, priorsVarName = NA, depth = 1,
  alpha = 1, usePriorEveryRecurse = FALSE)
}
\arguments{
\item{data, }{tidied data}

\item{groupVarName, }{grouping variable for subsetting DF}

\item{quantityVarName, }{entity name we're quantifying over
(i.e. "stars" in Peloquin & Frank (2016))}

\item{semanticsVarName, }{literal listener semantic values for RSA computations}

\item{itemVarName, }{unique items were comparing, probaby words
(i.e. "degrees" in Peloquin & Frank (2016))}

\item{costsVarName, }{costs variable name}

\item{priorsVarName, }{priors variable name}

\item{depth, }{recursive depth parameter}

\item{alpha, }{decision noise parameter level}

\item{usePriorEveryRecurse, }{boolean incorporate priors during each full recursion}
}
\value{
data frame with posterior predictions 'preds' column appended
}
\description{
Run RSA on tidied data frame with multiple groups
rsa-ready, tidied data should have a unique semantic value for each quantity * item combination.
(see formatting of peloquinFrank_2Alts data).
}
\examples{
rsa.runDf_grouped(data=peloquinFrank_2Alts,
          groupVarName="scale",
          quantityVarName="stars",
          itemVarName="words",
          semanticsVarName="speaker.p")

}
\keyword{(rsa.runDf_subset)}
\keyword{for}
\keyword{functionality}
\keyword{primary}
\keyword{run}
\keyword{wrapper}

