% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{unnameRSACols}
\alias{unnameRSACols}
\title{Return to original names (before renameRSACols()) for output}
\usage{
unnameRSACols(df, originalLabels)
}
\arguments{
\item{df, }{data frame with renamed columns}

\item{oldNames, }{original names}
}
\value{
, fill this out
}
\description{
Return to original names (before renameRSACols()) for output
}
\examples{
d <- data.frame(scales = rep("some_all", 10),
stars = as.factor(rep(1:5, 2)),
degrees = c(rep("strong", 5), rep("weak", 5)),
speaker.p = c(0, 0, 0, 0.3, 0.7,
0, 0.1, 0.15, 0.35, 0.40),
pragmatics = c(0, 0, 0, 0.15, 0.85,
0, 0.1, 0.25, 0.5, 0.15))
newDf <- renameRSACols(d, group = "scales", quantity = "stars",
item = "degrees", semantics = "speaker.p")
unnameRSACols(newDf$data, newDf$labels)

}
\seealso{
Called by \code{run_rrrsa}
}
\keyword{data_org}

