% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{utility}
\alias{utility}
\title{RSA utility}
\usage{
utility(items, costs = rep(0, length(items)), alpha = 1)
}
\arguments{
\item{items, }{literal semantic input vector <m_u1, m_u2, ...>
or <m1_u, m2_u,...>}

\item{alpha, }{decision nose parameter (see 'informativity()')}

\item{cost, }{cost vector <u1, u2, ...>}
}
\value{
, fill this out
}
\description{
Return normalized utility for a vector
}

