% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validateDims}
\alias{validateDims}
\title{Validate data entries}
\usage{
validateDims(m1, m2)
}
\arguments{
\item{m1, }{matrix 1}

\item{m2, }{matrix 2}
}
\value{
, fill this out
}
\description{
Compare dimensions of two matrices or vectors
}
\examples{
m1 <- matrix(data = 1:4, nrow = 2)
m2 <- matrix(data = 5:8, nrow = 2)
m3 <- matrix(data = 5:10, nrow = 2)
validateDims(m1, m2)
validateDims(m1, m3)

}
\keyword{safety_checks}

