% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_2Alts.R
\docType{data}
\name{peloquinFrank_2Alts}
\alias{peloquinFrank_2Alts}
\title{Two alternative model data from Peloquin and Frank 2016}
\description{
Two alternative model data from Peloquin and Frank 2016
}
\keyword{data}

