% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_4Alts.R
\docType{data}
\name{peloquinFrank_4Alts}
\alias{peloquinFrank_4Alts}
\title{Four alternatives data from Peloquin and Frank 2016}
\description{
Four alternatives data from Peloquin and Frank 2016
}
\keyword{data}

