% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_5Alts.R
\docType{data}
\name{peloquinFrank_5Alts}
\alias{peloquinFrank_5Alts}
\title{Five alternatives data from Peloquin and Frank 2016}
\description{
This data set contains empirical literal listener semantics
and pragmatic judgments for five alternatives scalar items from
Peloquin & Frank 2016.
}
\details{
1) quantityVarName :: "stars"

2) semanticsVarName :: "speaker.p"

3) itemVarName :: "words"

4) groupVarName :: "scale"

other) experiment number :: "exp"

other) Pragmatic judgments study 1 (e6), study 2 (e11)
}
\references{
\url{https://github.com/langcog/scalar_implicature}
}
\keyword{data}

