% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peloquinFrank_5Alts.R
\docType{data}
\name{peloquinFrank_5Alts}
\alias{peloquinFrank_5Alts}
\title{Five alternatives data from Peloquin and Frank 2016}
\description{
Five alternatives data from Peloquin and Frank 2016
}
\keyword{data}

