% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.convertVecType}
\alias{rsa.convertVecType}
\title{Convert vector data type}
\usage{
rsa.convertVecType(vec1, vec2)
}
\arguments{
\item{vec1, }{first vector with type we want to match}

\item{vec2, }{second vector with type we'd like to transform}
}
\value{
vec2 with type matched to vec1
}
\description{
Convert vector data type
}
\examples{
v1 <- as.factor(c("one", "two", "three"))
v2 <- c("one", "two", "three")
v3 <- c("1", "2", "3")
rsa.convertVecType(v1, v2)
rsa.convertVecType(v1, v3)

}

