% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.normVec}
\alias{rsa.normVec}
\title{Normalize vectors}
\usage{
rsa.normVec(v)
}
\arguments{
\item{v, }{vector to be normalized}
}
\value{
, normalized vector
}
\description{
Return a normalized vector
}
\keyword{data}
\keyword{processing}

