% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.reason}
\alias{rsa.reason}
\title{Run (multiple) iterations RSA \code{rsa.fullRecursion()}}
\usage{
rsa.reason(m, costs = m - m, priors = rep(1, nrow(m)), depth = 1,
  alpha = 1)
}
\arguments{
\item{m, }{matrix of semantics}

\item{costs, }{cost matrix dimensions identical to semantic matrix}

\item{priors, }{prior vector same length as nrow(m)}

\item{depth, }{number of recursions}

\item{alpha, }{decision hyper-param}
}
\value{
, fill this out
}
\description{
Return matrix after undergoing 'depth' recursions
calling 'fn' for computation
}
\examples{
m <- matrix(data = c(0, 0.2, 0.25, 0.25, 0.3, 0, 0, 0, 0.3, 0.7), nrow = 5)
rownames(m) <- 1:5
colnames(m) <- c("item1", "item2")
rsa.reason(m, 0)
rsa.reason(m, 2)
priors <- c(0.1, 0.1, 0.1, 0.1, 0.6)
}
\keyword{recursion}

