% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.renameCol}
\alias{rsa.renameCol}
\title{Rename df columns}
\usage{
rsa.renameCol(df, currNames, replacements)
}
\arguments{
\item{df, }{data frame to change col names}

\item{currNames, }{names we want to replace (passed as characters)}

\item{replacements, }{values used as replacements}
}
\value{
data frame with changed col names
}
\description{
return a data frame with columns renamed
(if they exist in the df passed in)
}
\keyword{"UPPER_CASE"))}
\keyword{"upper_case"),}
\keyword{<-}
\keyword{=}
\keyword{LETTERS[1:3])}
\keyword{c("LOWER_CASE",}
\keyword{c("lower_case",}
\keyword{data}
\keyword{data.frame(LOWER_CASE}
\keyword{df}
\keyword{letters[1:3],}
\keyword{processing}
\keyword{rsa.renameCol(df,}
\keyword{upper_case}

