% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rsa.runDf}
\alias{rsa.runDf}
\title{Run RSA on a tidied data frame (assumes one group)
--------------------------------------------------}
\usage{
rsa.runDf(data, quantityVarName, semanticsVarName, itemVarName,
  costsVarName = NA, priorsVarName = NA)
}
\arguments{
\item{data, }{tidied data}

\item{itemVarName, }{unique items were compring, probaby words
(i.e. "degrees" in Peloquin & Frank (2016))}

\item{costsVarName, }{costs variable name}

\item{priorsVarName, }{priors variable name}

\item{quanityVarName, }{entity name we're quantifying over
(i.e. "stars" in Peloquin & Frank (2016))}
}
\value{
, return data frame with 'pred' col appended
}
\description{
Expects tidied data with three required and two optional fields
1) quanityVarName         :: entity name we're quantifying over
2) itemVarName            :: unique items were considering,
3) semanticsVarName       :: literal listener semantics
4) optional costsVarName  :: costs
5) optional priorsVarName :: priors
}
\examples{

}
\keyword{functionality}
\keyword{primary}
\keyword{run}

