% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reasoning.R
\name{rsa.tuneDepthAlpha}
\alias{rsa.tuneDepthAlpha}
\title{Tune depth and alpha hyperparamters}
\usage{
rsa.tuneDepthAlpha(data, quantityVarName, semanticsVarName, itemVarName,
  groupName = NA, compareDataName, costsVarName = NA, priorsVarName = NA,
  depths = 1, alphas = 1, compareItems = NA)
}
\arguments{
\item{data, }{tidied data}

\item{semanticsVarName, }{semantic values for inference computation}

\item{itemVarName, }{unique items were comparing, probaby words}

\item{groupName, }{grouping variable if we have one}

\item{compareDataName, }{pragmatic judgments we're comparing to}

\item{costsVarName, }{costs variable name}

\item{priorsVarName, }{priors variable name}

\item{depths, }{vector of depths (in integers) for tuning}

\item{alphas, }{vector of alphas for tuning}

\item{compareItems, }{specific items (in itemVarName col) for data subsetting}

\item{quanityVarName, }{entity name we're quantifying over}
}
\value{
, return a list of \# alphas \times \# depths tuples with (r, depth, alpha)

return a matrix of posteriors
}
\description{
Return a list with number of alpha \times depths elements
each element includes a tuple of (correlation, alpha, depth)
}
\examples{
print("need to do examples and validation checks")
}
\keyword{data}
\keyword{recursion}
\keyword{tuning}

